+++ AlgorithmComparison/src/main/java/algorithms/algorithmcomparison/algorithms/SortAlgorithm.java
@@ -3,7 +3,7 @@ package algorithms.algorithmcomparison.algorithms;
 public interface SortAlgorithm {
 
 	public void sort(final int[] array);
-	public int getNumberOfKeyComparisons();
+	public long getNumberOfKeyComparisons();
 	public long getRuntimeMilliseconds();
 	public void reset();
 	public int[] getResultArray();
